backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
azure-identity
azure-ai-projects
azure-ai-inference
azure-cognitiveservices-speech
azure-storage-blob
azure-keyvault-secrets
openai
python-jose[cryptography]
python-multipart
python-dotenv
httpx
pydantic-settings
alembic
psycopg2-binary
redis
celery
neo4j
qdrant-client
scikit-learn
hdbscan
networkx
pandas
numpy
aiosqlite
